Formularios, Cloudinary.
- Cloudinary: Se crea nuevo componente para gestionar el cloudinary.
- Formularios: Se modifica componente newSector y newAgent para que gestionen la imagen mediante el componente cloudinary.
